requirements-runtime.txt
fastapi>=0.110
uvicorn>=0.29
asyncpg>=0.29
aiohttp>=3.9
python-dotenv>=1.0
pydantic>=2.7
web3>=6.18
openai>=1.30
prometheus-client>=0.19
unleashclient>=5.4
opentelemetry-api>=1.25
opentelemetry-sdk>=1.25
typer[all]>=0.12
tiktoken>=0.6
opencv-python-headless>=4.10      # «headless», чтобы не тянуть GUI
pydantic-settings>=2.1

alembic~=1.13
alembic~=1.13
alembic~=1.13
psycopg2-binary~=2.9
python-telegram-bot==21.0.1

